Remove character from string
A computer program receives a word and an unsigned integer less than length of the entered word. If unsigned integer does not meet the condition, reentry for the unsigned integer is requested until condition is met. Character at the entered integer is removed from the entered word and saved in another string. Saved string is displayed.